+++ protocols/kad/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Make `Kademlia::add_address` return `Result`.
   See [PR 4280](https://github.com/libp2p/rust-libp2p/pull/4280).
+- Emit `ModeChanged` event whenever we automatically reconfigure the mode.
+  See [PR 4503](https://github.com/libp2p/rust-libp2p/pull/4503).
 
 ## 0.44.6